chore: rename package autostart/lifecycle to lifecycle/executor
